final src

+++ src/MarioGame/core/src/com/nhs/game/Object/Enermy/Goomba.java
@@ -11,6 +11,8 @@ import com.badlogic.gdx.physics.box2d.CircleShape;
 import com.badlogic.gdx.physics.box2d.FixtureDef;
 import com.badlogic.gdx.physics.box2d.PolygonShape;
 import com.badlogic.gdx.utils.Array;
+import com.nhs.game.Object.Effect.EffectDef;
+import com.nhs.game.Object.Effect.ScoreText;
 import com.nhs.game.Object.Player.Mario;
 import com.nhs.game.Screens.ScreenManagement;
 import com.nhs.game.mariobros;
@@ -23,7 +25,7 @@ import static com.nhs.game.Global.global.FIREBALL_BIT;
 import static com.nhs.game.Global.global.GROUND_BIT;
 import static com.nhs.game.Global.global.ENERMY_BIT;
 import static com.nhs.game.Global.global.MARIO_BIT;
-import static com.nhs.game.Global.global.OBJECT_BIT;
+import static com.nhs.game.Global.global.PIPE_BIT;
 import static com.nhs.game.Global.global.PPM;
 import static com.nhs.game.UiManager.Hud.UpdateScore;
 
@@ -36,13 +38,11 @@ public class Goomba extends Enermy
     private boolean setDestroy;
     public Goomba(ScreenManagement screen, float x, float y) {
         super(screen, x, y);
-
         frames=new Array<TextureRegion>();
         for (int i=0;i<2;i++)
         {
             frames.add(new TextureRegion(screen.getAtlas().findRegion("goomba"),i*16,0,16,16));
         }
-
         wallAnimation=new Animation(0.4f,frames);
         stateTime=0;
         setBounds(getX(),getY(),16/PPM,16/PPM);
@@ -54,28 +54,30 @@ public class Goomba extends Enermy
 
     public void update(float dt)
     {
-        stateTime+=dt;
-        if (setDestroy && !eDestroyed)
-        {
-            world.destroyBody(b2body);
-            eDestroyed=true;
-            setRegion(new TextureRegion(screen.getAtlas().findRegion("goomba"),32,0,16,16));
-            stateTime=0;
+        if (eDestroyed) return;
 
-        } else if (!eDestroyed)
-        {
+        else
+            {
+                stateTime+=dt;
+                if (setDestroy && !eDestroyed)
+                {
+                    setRegion(new TextureRegion(screen.getAtlas().findRegion("goomba"),32,0,16,16));
+                    b2body.setUserData(null);
+                    world.destroyBody(b2body);
+                    b2body=null;
+                    eDestroyed=true;
+                    stateTime=0;
+                    return;
 
-            setPosition(b2body.getPosition().x-getWidth()/2,b2body.getPosition().y-getHeight()/2);
+                } else if (!eDestroyed)
+                {
+                    setPosition(b2body.getPosition().x-getWidth()/2,b2body.getPosition().y-getHeight()/2);
+                    setRegion( (TextureRegion)wallAnimation.getKeyFrame(stateTime,true));
+                    velocity.y=b2body.getLinearVelocity().y;
+                    b2body.setLinearVelocity(velocity);
 
-            setRegion( (TextureRegion)wallAnimation.getKeyFrame(stateTime,true));
-
-            velocity.y=b2body.getLinearVelocity().y;
-            b2body.setLinearVelocity(velocity);
-
-        }
-
-
-
+                }
+            }
 
     }
 
@@ -95,7 +97,7 @@ public class Goomba extends Enermy
         // category để nhận biết đó là object nào
         // mask là các object và object đang xét có thể va chạm
         fdef.filter.categoryBits=ENERMY_BIT;
-        fdef.filter.maskBits= GROUND_BIT |MARIO_BIT| COINS_BIT |BRICK_BIT|OBJECT_BIT|ENERMY_BIT|FIREBALL_BIT;
+        fdef.filter.maskBits= GROUND_BIT |MARIO_BIT| COINS_BIT |BRICK_BIT| PIPE_BIT |ENERMY_BIT|FIREBALL_BIT;
 
         fdef.shape=shape;
 
@@ -118,7 +120,7 @@ public class Goomba extends Enermy
         //set mask de goi lai trong collision
         b2body.createFixture(fdef).setUserData(this);
 
-
+        shape.dispose();
 
 
 
@@ -140,6 +142,8 @@ public class Goomba extends Enermy
 
     @Override
     public void hitOnHead(Mario mario) {
+        (screen).spawnEffect(new EffectDef(new Vector2(b2body.getPosition().x,b2body.getPosition().y+16/PPM),
+                ScoreText.class));
         //xóa b2body của goomba để k xét va chạm nữa
         setDestroy=true;
         UpdateScore(100);
@@ -162,6 +166,8 @@ public class Goomba extends Enermy
 
     @Override
     public void killEnermy() {
+        (screen).spawnEffect(new EffectDef(new Vector2(b2body.getPosition().x,b2body.getPosition().y+16/PPM),
+                ScoreText.class));
         setDestroy=true;
         UpdateScore(100);
         mariobros.manager.get("audio/sounds/stomp.wav",Sound.class).play();
